requirements/dev.txt
# requirements/dev.txt
-r common.txt

WebTest==2.0.15
django-webtest==1.7.7
django-debug-toolbar==1.2
